+++ lib/monad.dart
@@ -1,0 +1,3 @@
+/// A collection of functional programming structures.
+export 'src/monad/either.dart' show Either;
+export 'src/monad/optional.dart' show Optional;